# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Game On Tabletop API
  description: REST API for Game On Tabletop — a platform for organising tabletop
    game nights, tracking game collections, scheduling events, and managing group
    RSVPs.
  version: 1.0.0
paths:
...
  /api/series/{series_id}/occurrences:
    get:
      tags:
      - series
      summary: List occurrences for a series
      operationId: list_occurrences_api_series__series_id__occurrences_get
      security:
      - HTTPBearer: []
      parameters:
      - name: series_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Series Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OccurrenceResponse'
                title: Response List Occurrences Api Series  Series Id  Occurrences
                  Get
        '401':
          description: Unauthorized
        '403':
          description: Not a member
        '404':
          description: Series not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - occurrences
      summary: Create a one-off occurrence for a series (organiser only)
      operationId: create_occurrence_api_series__series_id__occurrences_post
      security:
      - HTTPBearer: []
      parameters:
      - name: series_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Series Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OccurrenceCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OccurrenceResponse'
        '401':
          description: Unauthorized
        '403':
          description: Not an organiser
        '404':
          description: Series not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/me/occurrences:
    get:
      tags:
      - occurrences
      summary: List upcoming (or past) occurrences across all the user's groups
      operationId: list_my_occurrences_api_me_occurrences_get
      security:
      - HTTPBearer: []
      parameters:
      - name: past
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Past
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MyOccurrenceItem'
                title: Response List My Occurrences Api Me Occurrences Get
        '401':
          description: Unauthorized
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/occurrences/{occurrence_id}:
    get:
      tags:
      - occurrences
      summary: Get a single occurrence with RSVP counts
      operationId: get_occurrence_api_occurrences__occurrence_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: occurrence_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Occurrence Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OccurrenceResponse'
        '401':
          description: Unauthorized
        '403':
          description: Not a member of this group
        '404':
          description: Occurrence not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - occurrences
      summary: Update an occurrence (organiser only)
      operationId: update_occurrence_api_occurrences__occurrence_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: occurrence_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Occurrence Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OccurrenceUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OccurrenceResponse'
        '401':
          description: Unauthorized
        '403':
          description: Not an organiser
        '404':
          description: Occurrence not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/occurrences/{occurrence_id}/rsvp:
    post:
      tags:
      - occurrences
      summary: Submit or update an RSVP for an occurrence
      operationId: upsert_rsvp_api_occurrences__occurrence_id__rsvp_post
      security:
      - HTTPBearer: []
      parameters:
      - name: occurrence_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Occurrence Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RsvpCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RsvpOut'
        '400':
          description: Cannot RSVP to a cancelled occurrence
        '401':
          description: Unauthorized
        '403':
          description: Not a member of this group
        '404':
          description: Occurrence not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/occurrences/{occurrence_id}/rsvps:
    get:
      tags:
      - occurrences
      summary: List RSVPs for an occurrence
      operationId: list_rsvps_api_occurrences__occurrence_id__rsvps_get
      security:
      - HTTPBearer: []
      parameters:
      - name: occurrence_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Occurrence Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RsvpOut'
                title: Response List Rsvps Api Occurrences  Occurrence Id  Rsvps Get
        '401':
          description: Unauthorized
        '403':
          description: Not a member of this group
        '404':
          description: Occurrence not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/occurrences/{occurrence_id}/photos:
    get:
      tags:
      - occurrences
      summary: List photos uploaded to an occurrence
      operationId: list_photos_api_occurrences__occurrence_id__photos_get
      security:
      - HTTPBearer: []
      parameters:
      - name: occurrence_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Occurrence Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OccurrencePhotoOut'
                title: Response List Photos Api Occurrences  Occurrence Id  Photos
                  Get
        '401':
          description: Unauthorized
        '403':
          description: Not a member of this group
        '404':
          description: Occurrence not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - occurrences
      summary: Upload a photo to an occurrence
      operationId: upload_photo_api_occurrences__occurrence_id__photos_post
      security:
      - HTTPBearer: []
      parameters:
      - name: occurrence_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Occurrence Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_photo_api_occurrences__occurrence_id__photos_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OccurrencePhotoOut'
        '400':
          description: Invalid file type or file too large
        '401':
          description: Unauthorized
        '403':
          description: Not a member of this group
        '404':
          description: Occurrence not found
        '502':
          description: Storage upload failed or not configured
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/occurrences/{occurrence_id}/photos/{photo_id}:
    delete:
      tags:
      - occurrences
      summary: Delete a photo from an occurrence (uploader or organiser only)
      operationId: delete_photo_api_occurrences__occurrence_id__photos__photo_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: occurrence_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Occurrence Id
      - name: photo_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Photo Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: Unauthorized
        '403':
          description: Cannot delete this photo
        '404':
          description: Photo not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_photo_api_occurrences__occurrence_id__photos_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_upload_photo_api_occurrences__occurrence_id__photos_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MyOccurrenceItem:
      properties:
        id:
          type: string
          title: Id
        occurrence_date:
          type: string
          title: Occurrence Date
        start_time:
          type: string
          title: Start Time
        end_time:
          anyOf:
          - type: string
          - type: 'null'
          title: End Time
        status:
          type: string
          title: Status
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        series_id:
          type: string
          title: Series Id
        series_title:
          type: string
          title: Series Title
        group_id:
          type: string
          title: Group Id
        group_name:
          type: string
          title: Group Name
        my_rsvp:
          anyOf:
          - type: string
          - type: 'null'
          title: My Rsvp
      type: object
      required:
      - id
      - occurrence_date
      - start_time
      - end_time
      - status
      - notes
      - series_id
      - series_title
      - group_id
      - group_name
      - my_rsvp
      title: MyOccurrenceItem
    OccurrenceCreate:
      properties:
        occurrence_date:
          type: string
          format: date
          title: Occurrence Date
        start_time:
          type: string
          format: time
          title: Start Time
        end_time:
          anyOf:
          - type: string
            format: time
          - type: 'null'
          title: End Time
        location_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Location Id
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        rsvp_deadline:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Rsvp Deadline
      type: object
      required:
      - occurrence_date
      - start_time
      title: OccurrenceCreate
    OccurrencePhotoOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        occurrence_id:
          type: string
          format: uuid
          title: Occurrence Id
        uploaded_by:
          type: string
          format: uuid
          title: Uploaded By
        photo_url:
          type: string
          title: Photo Url
        caption:
          anyOf:
          - type: string
          - type: 'null'
          title: Caption
        created_at:
          type: string
          format: date-time
          title: Created At
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
      type: object
      required:
      - id
      - occurrence_id
      - uploaded_by
      - photo_url
      - caption
      - created_at
      title: OccurrencePhotoOut
    OccurrenceResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        series_id:
          type: string
          format: uuid
          title: Series Id
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        occurrence_date:
          type: string
          format: date
          title: Occurrence Date
        start_time:
          type: string
          format: time
          title: Start Time
        end_time:
          anyOf:
          - type: string
            format: time
          - type: 'null'
          title: End Time
        location_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Location Id
        status:
          type: string
          title: Status
        rsvp_deadline:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Rsvp Deadline
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        postponed_from_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Postponed From Date
        is_auto_generated:
          type: boolean
          title: Is Auto Generated
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        rsvp_counts:
          anyOf:
          - additionalProperties:
              type: integer
            type: object
          - type: 'null'
          title: Rsvp Counts
      type: object
      required:
      - id
      - series_id
      - title
      - occurrence_date
      - start_time
      - end_time
      - location_id
      - status
      - rsvp_deadline
      - notes
      - postponed_from_date
      - is_auto_generated
      - created_at
      - updated_at
      title: OccurrenceResponse
    OccurrenceUpdate:
      properties:
        status:
          anyOf:
          - type: string
            enum:
            - scheduled
            - postponed
            - cancelled
          - type: 'null'
          title: Status
        occurrence_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Occurrence Date
        start_time:
          anyOf:
          - type: string
            format: time
          - type: 'null'
          title: Start Time
        end_time:
          anyOf:
          - type: string
            format: time
          - type: 'null'
          title: End Time
        location_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Location Id
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        rsvp_deadline:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Rsvp Deadline
        postponed_from_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Postponed From Date
      type: object
      title: OccurrenceUpdate
...
    RsvpCreate:
      properties:
        response:
          type: string
          enum:
          - 'yes'
          - 'no'
          - maybe
          title: Response
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - response
      title: RsvpCreate
    RsvpOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        occurrence_id:
          type: string
          format: uuid
          title: Occurrence Id
        user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: User Id
        response:
          type: string
          title: Response
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        responded_at:
          type: string
          format: date-time
          title: Responded At
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
      type: object
      required:
      - id
      - occurrence_id
      - user_id
      - response
      - note
      - responded_at
      title: RsvpOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer